T5909: Use warning for attempt to connect to containers registry
This fixes the issue when system does not have the Internet
connection while attempting to use container registry with login
and password. It causes after rebooting we do not have section
"container" at all.

The container priority 450
The protocols static priority 480

The best way to fix is using `Warning` insted of `raise ConfigError`