wireguard: T5924: harden migration script logic 22-to-23
The original commit 2c1c36135 ("wireguard: T5413: Blocked adding the peer with
the router's public key") did not honor the fact that there might be no
private-key CLI node defined for a WireGuard interface. If this is the case,
private-key defaults to "default". This fact needs to be handled in the
migration script.