ddclient: T5966: Streamline dynamic dns op-mode configuration
Update op-mode for dynamic dns to standardize on `vyos.opmode`. All
methods of `op_mode/dns_dynamic.py` are now available in standardized
`op_mode/dns.py`.

Move op-mode command `update dns dynamic` to `reset dns dynamic` to
reflect that it is not an update but a reset of the dynamic dns service.

Also, make the help texts more consistent for all op-mode commands for
`dns dynamic` and `dns forwarding`.