container: T5909: move registry login to op-mode
It does not make sense to perform the "podman login" command when setting up
containers, as images are not automatically pulled in from the registry - due
to issues with the default route during startup.

The same issue manifests in "podman login" where we can not login to a registry
unless there is a default route present.

This commit changes the behavior that the container registry is part of the
configuration, but it is only referenced during "add container image" and thus
never during system boot.

(cherry picked from commit baf30d8319ef4d0f0cc4cdf0f7c12f03f8a492b6)