vyos.configverify: T6131: verify_interface_exists() checks CLI interfaces, too
Extend the way how we determine if interfaces exist in VyOS. In the past we
only validated if the interface in question really exists at the OS level.
This has some drawbacks as services (like OSPF or OSPFv3) can also handle
interfaces dynamically which appear or leaf the OS.

This commit not only checks for OS interfaces but also if the interface in
question was configured at the CLI level, this is proof enough to pass the
check. If it does not exist at the CLI level, we continue searching it it's
maybe a Kernel interface - useful for container networks.

In addition we can now not only raise() an error but simply show a warning if
an interface does not exist.

(cherry picked from commit f7250ecf1d119f14d72f99ee379deaaae0790f0e)