dhcpv6-client: T2590: fix vyos-hostsd update for nameserver and search domains
After migrating from ISC DHCLIENT for IPv6 to wide-dhcp-client the logic which
was present to update /etc/resolv.conf with the DHCP specified nameservers and
also the search domain list was no longer present.

This commit adds a per interface rendered script to inform vyos-hostsd about
the received IPv6 nameservers and search domains.