container: T6208: rename "cap-add" CLI node to "capability"
Containers have the ability to add Linux system capabilities to them, this is
done using the "set container name <name> cap-add" command.

The CLI node sounds off and rather should be "set container name <name>
capability" instead as we use and pass a capability to a container and not
add/invent new ones.