T6716: don't automatically set ethernet offload
Remove the lines of code that checked if the kernel had offloading
enabled and was then forcing the config to set it to "on." The
behavior now mirrors the config and offloading will only be enabled
if the config is explicitly set to enabled.

Note: the code is still present to disable the offloading, in the
config, if the kernel doesn't support it.

Note(2): Allow the previous behavior where the offload settings get set,
based on the Kernel, if the boot is a live boot.