Enable IPv6 on wwan0 (#5)
VyOS sets addr_gen_mode to 1 (do not generate link-local) for *ALL*
interfaces in case it is going to be added to a bridge, in which case
it must not have any addresses.  However, cellular gets these from the
provider and hence can never be in a bridge so this is not approriate.

Normally this would be done via NetworkManager but VyOS does not use that.
So we do it via a sysctl overriding their existing 31-vyos-addr_gen_mode.conf
but for the wwan0 interface only.