Add python3-systemd dependency (for x86_64 build).
Needed for vyos-vpp which is part of the vyos-1x build.

This was formerly handled by adding the dependency to
nexus-build/updates/arm64fs.toml but this does not help
for a non-arm build (e.g. x86).  The actual dependent
is vyos-1x via vyos-vpp/python/vyos/vpp/control_vpp.py
so this is the right place.

This change does not affect arm builds, but it does mean that
the python3-systemd dependency in arm64fs.toml is no longer
needed and will be removed at some point.