Support to disable/block IPv6
This implements the new "set firewall ipv6 block" configuration command
which effectively works like the old "set system ipv6 disable".

VyOS once had a "set system ipv6 disable" but it was deprecated in 1.3
and removed completely in 1.4 with no intent to replace it in any way:

    https://vyos.dev/T4346

Their original implementation disabled it on the kernel level,
but per their comments this causes all sorts of other problems:

    https://vyos.dev/T4319

We do not want to use the same command to avoid confusion with the
old command which worked differently.  Also, VyOS migration scripts
will remove any such command since it is deprecated.

This PR effectively blocks IPv6 via configuration instead, via a
single new setting "set firewall ipv6 block". This in effect makes VyOS
behave *as if* you did "ipv6 address no-default-link-local" and "ipv6
disable-forwarding" on each interface, plus adds firewall rules to drop
IPv6 packets for input/output/forward.

It is important to note that this is just a single configuration setting.
It does not create additional settings; the system will act as if
they were set as a consequence.

Note that since the ipv6 address and forwarding settings are normally
off, it will not be a problem if the user already enables them,
i.e. it will cause no configuration conflicts.

If the user has other ipv6 firewall rules, they will get a warning message
that that "firewall ipv6 block" is in place which overrides their settings.

There is also a message that a reboot is required when ipv6 block changes,
which is how it was with the previous VyOS implementation and is how
it works on IRG/SCRx already.

Here is the output after a "commit" of using the new feature, where
there also happened to be a conflicting ipv6 input filter setting in place.

    Warning: "firewall ipv6 input filter default-action" forced to "drop" because "firewall ipv6 block" is set
    Changing firewall ipv6 block will take effect only when the system is rebooted.